Add storage support; Refactor uploading services
