+++ BUILD
@@ -4,8 +4,11 @@ load("@bazel_tools//tools/python:toolchain.bzl", "py_runtime_pair")
 py_binary(
     name = "benchmark",
     srcs = ["benchmark.py"],
+    python_version = "PY2",
     deps = [
         "//utils",
+        "//utils:bigquery_upload",
+        "//utils:storage_upload",
         requirement("absl-py"),
         requirement("GitPython"),
         requirement("gitdb2"),
